Add files via upload
Initially, the Database was created, and values were inserted. For the database, we used the tables from a previous submission. New data was inserted into the database. These data entries have simple names and numbers, with usernames following the format "userX," where X is a number.

Additionally, the necessary consistency between the tables was maintained, ensuring that jobs require the appropriate qualifications and that degrees align with those qualifications. Users are registered in ascending numerical order across all entries:

user1 to user18 are registered as employees.
user19 to user36 are registered as evaluators.
Later, user37 to user40 were added as the Database Administrators.